Fix: change sequence of showing level up dialog and change django testcode
