+++ android/RunUsAndroid/app/src/main/java/com/example/runusandroid/RetrofitClient.java
@@ -12,9 +12,9 @@ import retrofit2.converter.gson.GsonConverterFactory;
 public class RetrofitClient {
 
     // 아래 코드를 수정 후 커밋하지 마시오!!
-    private final static String BASE_URL = "http://ec2-3-36-116-64.ap-northeast-2.compute.amazonaws.com:3000";
+    //private final static String BASE_URL = "http://ec2-3-36-116-64.ap-northeast-2.compute.amazonaws.com:3000";
     // 위 코드를 수정 후 커밋하지 마시오!!
-    //private final static String BASE_URL = "http://10.0.2.2:8000";
+    private final static String BASE_URL = "http://10.0.2.2:8000";
 
     private static String authToken = null;
     private static Retrofit retrofit = null;
