+++ android/SocketServer/.gitignore
@@ -3,6 +3,8 @@
 # Compiled class file
 *.class
 
+out/
+
 # Log file
 *.log
